+++ Source/SelfService/Web/microservice/base/view.tsx
@@ -114,7 +114,12 @@ export const View: React.FunctionComponent<Props> = (props) => {
         const runtimeImage = currentMicroservice.live.images.find(img => img.name === 'runtime')?.image
             || 'n/a';
 
+        const headCommand = {
+            command: [],
+            args: []
+        };
 
+        const environmentInfo = application.environments.find(_environment => _environment.name === environment)!;
         // TODO currently we don't use the ms.extra.ingress in the view
         // Look to "liveIngressView" for how we "set" the data to uniq paths
         ms = {
@@ -135,6 +140,8 @@ export const View: React.FunctionComponent<Props> = (props) => {
                 isPublic: true,
                 headImage,
                 runtimeImage,
+                headCommand,
+                connections: environmentInfo.connections,
             },
         };
     }